feat: adding tempo
