+++ backend/app/main.py
@@ -1,7 +1,4 @@
-import os
-
 import app.services.firebase
-import sentry_sdk
 from app.api.v1.routers import (
     auth,
     auto_trade,
@@ -13,46 +10,19 @@ from app.api.v1.routers import (
     stock,
     worker,
 )
-from app.api.v1.routers.metrics import api_exceptions_total
 from app.cron.cron import start_autotrade_scheduler
 from app.middleware.cors import add_cors_middleware
 from app.middleware.firebase import add_firebase_auth_middleware
 from app.middleware.gzip import add_gzip_middleware
 from app.middleware.ratelimit import add_rate_limit_middleware
 from app.middleware.security import add_security_headers
-from fastapi import FastAPI, HTTPException, Request
-from fastapi.responses import JSONResponse
-
-sentry_sdk.init(
-    dsn=os.getenv("SENTRY_DSN"),
-    traces_sample_rate=0,
-    sample_rate=0.1,
-    environment=os.getenv("ENV", "production"),
-    send_default_pii=True,
-)
+from app.services.sentry import setup_sentry
+from app.services.tracing import setup_tracing
+from fastapi import FastAPI
 
 app = FastAPI()
-
-
-# Global Exception Handlers
-@app.exception_handler(Exception)
-async def global_exception_handler(request: Request, exc: Exception):
-    sentry_sdk.capture_exception(exc)
-    api_exceptions_total.inc()
-    return JSONResponse(
-        status_code=500,
-        content={"detail": "Internal server error"},
-    )
-
-
-@app.exception_handler(HTTPException)
-async def http_exception_handler(request: Request, exc: HTTPException):
-    if exc.status_code >= 500:
-        sentry_sdk.capture_exception(exc)
-    return JSONResponse(
-        status_code=exc.status_code,
-        content={"detail": exc.detail},
-    )
+setup_sentry(app)
+setup_tracing(app)
 
 
 # Scheduler
